model call

+++ test.py
@@ -1,51 +1,42 @@
 import streamlit as st
 import os
 import cv2
-import yaml
 import torch
 import numpy as np
 from PIL import Image
 from typing import Union
 from ultralytics import YOLO
-import yaml
 import io
+
+from ultralytics.utils.plotting import Annotator
 
 # Define the device to be used for computation
 DEVICE = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
 
 # Initialize YOLO model
-model = YOLO('best.pt')
-
-
-# Function for loading data from yaml file
-def load_yaml(file_path: str) -> dict:
-    """
-    Load YAML file.
-
-    Args:
-        file_path (str): Path to the YAML file.
-
-    Returns:
-        dict: Loaded YAML data.
-    """
-    with open(file_path, 'r') as file:
-        try:
-            data = yaml.safe_load(file)
-            return data
-        except yaml.YAMLError as exc:
-            print(exc)
+model = YOLO('best_no_transfer.pt')
 
 
 # Function for removing temporary files
 def remove_temp(temp_file: str = 'temp') -> None:
     """
-    Remove all files in the specified temporary directory.
+    Remove all files in the specified temporary directory. Creates the directory if it does not exist.
 
     Args:
         temp_file (str, optional): Path to the temporary directory. Defaults to 'temp'.
     """
+    # Check if the directory exists
+    if not os.path.exists(temp_file):
+        # Create the directory if it does not exist
+        os.makedirs(temp_file)
+        print(f"Directory '{temp_file}' was created")
+        return
+
+    # If the directory exists, proceed to remove files
     for file in os.listdir(temp_file):
         os.remove(os.path.join(temp_file, file))
+    print(f"All files in '{temp_file}' have been removed.")
+
 
 
 # Function for downloading an image with detected objects
@@ -111,7 +102,12 @@ def video_detect(uploaded_video: Union[None, io.BytesIO], confidence_threshold: 
         confidence_threshold (float): Confidence threshold for object detection.
         max_detections (int): Maximum number of detections.
     """
-    
+    fruits = []
+    class_names = ['Apple', 'Almond', 'Banana', 'Beetroot', 'Bitter_Gourd', 'Bottle_Gourd', 'Cabbage', 'Capsicum',
+                   'Carrot', 'Cauliflower', 'Cherry', 'Chilli', 'Coconut', 'Cucumber', 'EggPlant',
+                   'Ginger', 'Grape', 'Green_Orange', 'Kiwi', 'Maize', 'Mango', 'Melon',
+                   'Okra', 'Onion', 'Orange', 'Peach', 'Pear', 'Peas', 'Pineapple', 'Pomegranate',
+                   'Potato', 'Radish', 'Strawberry', 'Tomato', 'Turnip', 'Walnut', 'Watermelon']
     # Check if a video is uploaded
     if uploaded_video is not None:
         # Create a temporary file to save the uploaded video
@@ -119,7 +115,7 @@ def video_detect(uploaded_video: Union[None, io.BytesIO], confidence_threshold: 
 
         # Write uploaded video content to the temporary file
         with open(temp_video_path, "wb") as temp_video_file:
-            temp_video_file.write(uploaded_video.getbuffer())
+            temp_video_file.write(uploaded_video.read())
 
         # Open the uploaded video file
         cap = cv2.VideoCapture(temp_video_path)
@@ -132,24 +128,33 @@ def video_detect(uploaded_video: Union[None, io.BytesIO], confidence_threshold: 
             ret, frame = cap.read()
             if not ret:
                 break
-            
-            # Convert frame to PIL Image
-            img = Image.fromarray(frame)
+
 
             # Perform object detection - assuming your model has a similar API to Ultralytics YOLO
-            results = model.predict(img, conf=confidence_threshold, max_det=max_detections, device=DEVICE)
+            results = model.predict(frame, conf=confidence_threshold, max_det=max_detections, device=DEVICE)
 
-            # Visualization and conversion to NumPy for display - assuming results have .render() method
-            frame = np.squeeze(results.render())
+            for r in results:
 
-            # Update the frame in the Streamlit app
+                annotator = Annotator(frame)
+
+                boxes = r.boxes
+                for box in boxes:
+                    b = box.xyxy[0]  # get box coordinates in (left, top, right, bottom) format
+                    c = box.cls
+                    annotator.box_label(b, model.names[int(c)])
+                detection_count = r.boxes.shape[0]
+
+                for i in range(detection_count):
+                    cls = int(r.boxes.cls[i].item())
+                    name = r.names[cls]
+                    if name in class_names and name not in fruits:
+                        fruits.append(name)
+                        print(fruits)
             stframe.image(frame, channels="BGR", use_column_width=True)
 
         # Release the video capture object and remove the temp file
         cap.release()
         os.remove(temp_video_path)
-
-
 
 
 # Set Streamlit page configuration
@@ -182,7 +187,6 @@ elif source == "Video":
     uploaded_video = st.sidebar.file_uploader("Choose a video...", type=["mp4"])
 
 
-
 # Confidence threshold and max detections sliders
 confidence_threshold = st.sidebar.slider("Confidence Threshold", min_value=0.0, max_value=1.0, value=0.25, step=0.01)
 max_detections = st.sidebar.slider("Max Detections", min_value=1, max_value=500, value=300, step=1)
@@ -194,9 +198,9 @@ if uploaded_image is not None:
     image_detect(image=uploaded_image, confidence_threshold=confidence_threshold,
                  max_detections=max_detections)
 
-elif uploaded_video:
+elif uploaded_video is not None:
     # Object detection for uploaded video
-    video_detect(source='video', uploaded_video=uploaded_video, confidence_threshold=confidence_threshold,
+    video_detect(uploaded_video=uploaded_video, confidence_threshold=confidence_threshold,
                  max_detections=max_detections)
 
     # Remove temporary files
